complete generelization of using checker
