+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/checkers/nullaway/FixSerializationConfig.java
@@ -20,9 +20,21 @@
  * THE SOFTWARE.
  */
 
-package edu.ucr.cs.riple.core.util;
+package edu.ucr.cs.riple.core.checkers.nullaway;
 
+import java.io.File;
+import java.util.UUID;
 import javax.annotation.Nullable;
+import javax.xml.parsers.DocumentBuilder;
+import javax.xml.parsers.DocumentBuilderFactory;
+import javax.xml.parsers.ParserConfigurationException;
+import javax.xml.transform.Transformer;
+import javax.xml.transform.TransformerException;
+import javax.xml.transform.TransformerFactory;
+import javax.xml.transform.dom.DOMSource;
+import javax.xml.transform.stream.StreamResult;
+import org.w3c.dom.Document;
+import org.w3c.dom.Element;
 
 /**
  * IMPORTANT NOTE: THIS CLASS IS COPIED FROM NULLAWAY, WE COPIED THE CLASS CONTENT HERE TO REMOVE
@@ -97,7 +109,7 @@ public class FixSerializationConfig {
      */
     public void writeAsXML(String path) {
       FixSerializationConfig config = this.build();
-      Utility.writeNullAwayConfigInXMLFormat(config, path);
+      writeNullAwayConfigInXMLFormat(config, path);
     }
 
     public FixSerializationConfig build() {
@@ -105,6 +117,54 @@ public class FixSerializationConfig {
         throw new IllegalStateException("did not set mandatory output directory");
       }
       return new FixSerializationConfig(suggestEnabled, suggestEnclosing, fieldInitInfo, outputDir);
+    }
+
+    /**
+     * Writes the {@link FixSerializationConfig} in {@code XML} format.
+     *
+     * @param config Config file to write.
+     * @param path Path to write the config at.
+     */
+    public static void writeNullAwayConfigInXMLFormat(FixSerializationConfig config, String path) {
+      DocumentBuilderFactory docFactory = DocumentBuilderFactory.newInstance();
+      try {
+        DocumentBuilder docBuilder = docFactory.newDocumentBuilder();
+        Document doc = docBuilder.newDocument();
+
+        // Root
+        Element rootElement = doc.createElement("serialization");
+        doc.appendChild(rootElement);
+
+        // Suggest
+        Element suggestElement = doc.createElement("suggest");
+        suggestElement.setAttribute("active", String.valueOf(config.suggestEnabled));
+        suggestElement.setAttribute("enclosing", String.valueOf(config.suggestEnclosing));
+        rootElement.appendChild(suggestElement);
+
+        // Field Initialization
+        Element fieldInitInfoEnabled = doc.createElement("fieldInitInfo");
+        fieldInitInfoEnabled.setAttribute("active", String.valueOf(config.fieldInitInfoEnabled));
+        rootElement.appendChild(fieldInitInfoEnabled);
+
+        // Output dir
+        Element outputDir = doc.createElement("path");
+        outputDir.setTextContent(config.outputDirectory);
+        rootElement.appendChild(outputDir);
+
+        // UUID
+        Element uuid = doc.createElement("uuid");
+        uuid.setTextContent(UUID.randomUUID().toString());
+        rootElement.appendChild(uuid);
+
+        // Writings
+        TransformerFactory transformerFactory = TransformerFactory.newInstance();
+        Transformer transformer = transformerFactory.newTransformer();
+        DOMSource source = new DOMSource(doc);
+        StreamResult result = new StreamResult(new File(path));
+        transformer.transform(source, result);
+      } catch (ParserConfigurationException | TransformerException e) {
+        throw new RuntimeException("Error happened in writing config.", e);
+      }
     }
   }
 }
